requirements.txt
langchain>=0.1.0
langgraph>=0.0.50
langchain-google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pandas>=2.0.0
openpyxl>=3.1.0
markdown>=3.5.0
rich>=13.7.0
pydantic>=2.5.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
youtube-transcript-api>=0.6.0
aiohttp>=3.9.0
asyncio>=3.4.3
typer>=0.9.0
python-json-logger>=2.0.7
